+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
-- Removed `page.state.event` and store the event object in `page.event` property.
+- Changed the signature of the navigator handler.
+- Removed `page.state.event`.
 - Renamed `page.state.cache` to `page.state.html`.
 - Renamed extensions from `.jsm` to `.js` due `Content-Type` header issues
 - The `package.json` is now more browser-friendly using `browser` and `files` keys